+++ CHANGES.md
@@ -1,9 +1,11 @@
 ## [HEAD]
 
 - Disable browser history on Chrome iOS ([#146])
+- Add ES2015 module build ([#152])
 
 [HEAD]: https://github.com/rackt/history/compare/latest...HEAD
 [#146]: https://github.com/rackt/history/pull/146
+[#152]: https://github.com/rackt/history/pull/152
 
 ## [v1.13.1]
 > Nov 13, 2015