+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## [HEAD]
 
+#### New features
+- Accept objects in `history.push` and `history.replace` ([#141])
+
 #### Bug fixes
 - Disable browser history on Chrome iOS ([#146])
 - Do not convert same-path PUSH to REPLACE if the hash has changed ([#167]) 
@@ -9,6 +12,7 @@
 - Use query-string module instead of qs to save on bytes ([#121])
 
 [HEAD]: https://github.com/rackt/history/compare/latest...HEAD
+[#141]: https://github.com/rackt/history/pull/141
 [#146]: https://github.com/rackt/history/pull/146
 [#152]: https://github.com/rackt/history/pull/152
 [#167]: https://github.com/rackt/history/pull/167